Moved code from ast.ts to ast-search.ts
